Add not found handling and responses to Category routes. Update Category routes to respond with updated or deleted object.

+++ routes/api/category-routes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const { Category, Product } = require('../../models');
 
-
+// GET all Categories
 router.get('/', async (req, res) => {
   
   try {
@@ -18,6 +18,7 @@ router.get('/', async (req, res) => {
 
 });
 
+// GET Category by id
 router.get('/:id', async (req, res) => {
   
   try {
@@ -26,6 +27,10 @@ router.get('/:id', async (req, res) => {
         { model: Product }
       ]
     });
+    
+    if (!categoryData) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
     const category = categoryData.get({ plain: true });
     res.status(200).json(category);
   } catch (err) {
@@ -34,6 +39,7 @@ router.get('/:id', async (req, res) => {
 
 });
 
+// POST new Category 
 router.post('/', async (req, res) => {
   
   try {
@@ -45,20 +51,31 @@ router.post('/', async (req, res) => {
 
 });
 
-// TODO: doesn't return updated row
+// PUT update Category by id
 router.put('/:id', async (req, res) => {
   
   try {
-    const updatedCat = await Category.update(
+
+    const foundUpdatedCatData = await Category.findByPk(req.params.id);
+
+    if (!foundUpdatedCatData) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
+
+    const foundCat = foundUpdatedCatData.get({ plain: true });
+    
+    await Category.update(
       {
         category_name: req.body.category_name
       },
       {
-        where: { id: req.params.id },
-        returning: true,
-        plain: true
+        where: { id: foundCat.id }
       }
     );
+
+    const updatedCatData = await Category.findByPk(foundCat.id);
+    const updatedCat = updatedCatData.get({ plain: true });
+    
     res.status(200).json(updatedCat);
   } catch (err) {
     res.status(500).json(err);
@@ -66,16 +83,29 @@ router.put('/:id', async (req, res) => {
 
 });
 
-// TODO: doesn't return deleted Cat
+// DELETE Category by id. Deletes associated Products
 router.delete('/:id', async (req, res) => {
   
   try {
-    const deletedCat = await Category.destroy({ 
-      where: { id: req.params.id },
-      returning: true,
-      plain: true
+    const foundCatData = await Category.findByPk(req.params.id, {
+      include: [
+        { model: Product }
+      ]
     });
-    res.status(200).json(deletedCat);
+
+    if (!foundCatData) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
+    const foundCat = foundCatData.get({ plain: true });
+    
+    await foundCat.products.forEach(product => {
+      Product.destroy({ where: { id: product.id } });
+    });
+    
+    await Category.destroy({ 
+      where: { id: foundCat.id }
+    });
+    res.status(200).json(foundCat);
   } catch (err) {
     res.status(500).json(err);
   }
